Move Serializable trait to `traits` directory
