+++ src/traits/mod.rs
@@ -1,0 +1,3 @@
+mod serializable;
+
+pub use self::serializable::Serializable;
